+++ CHANGELOG.md
@@ -19,3 +19,5 @@
 - remove fadj(g, v) in favor of out_neighbors(g, v)
 - more efficient equality test for DiGraph
 - remove fadj(g) in favor of out_adjlist
+- add `graph` and `digraph`
+- add `edge`